restructures Carousel component file organization
