+++ src/components/ProjectsPage.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-import Carousel from './Carousel'
+import Carousel from './Carousel/Carousel'
 
 const ProjectsPage = () => (
   <div>
